### 2020-05-28 #2271 @whedon
[PRE REVIEW]: COVID-19 Data Hub
**Submitting author:** @ArdiaD (<a href="http://orcid.org/0000-0003-2823-782X">David Ardia</a>)
**Repository:** <a href="https://github.com/covid19datahub/COVID19/" target ="_blank">https://github.com/covid19datahub/COVID19/</a>
**Version:** 2.0.7

Thanks for submitting your paper to JOSS ArdiaD. **Currently, there isn't an JOSS editor assigned** to your paper.

ArdiaD if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
