### 2020-05-28 #2271 @ArdiaD
indeed I am a co-author ;)

Sent from my iPhone

> On 28 May 2020, at 19:43, Daniel S. Katz <notifications@github.com> wrote:
>
> I'm confused by this, since @ArdiaD is the author
>
