+++ pkgdown.yml
@@ -204,7 +204,7 @@ articles:
   ZAF: iso/ZAF.html
   ZMB: iso/ZMB.html
   ZWE: iso/ZWE.html
-last_built: 2020-05-28T20:28Z
+last_built: 2020-05-29T00:35Z
 urls:
   reference: https://covid19datahub.io/reference
   article: https://covid19datahub.io/articles
