+++ R/utils.R
@@ -1,56 +1,56 @@
 #' @importFrom dplyr %>%
 NULL
 
-.onAttach <- function(libname, pkgname) {
-
-  if(interactive() & requireNamespace('COVID19', quietly = TRUE)){
-
-    packageStartupMessage("The coronavirus situation is changing fast. Checking for updates...")
-
-    description <- try(readLines('https://raw.githubusercontent.com/covid19datahub/COVID19/master/DESCRIPTION'), silent = TRUE)
-    if(class(description)=="try-error")
-      return()
-      
-    id <- which(startsWith(prefix = "Version:", x = description))
-    v  <- as.package_version(gsub(pattern = "^Version:\\s*", replacement = "", x = description[id]))
-
-    if(v > utils::packageVersion(pkg = "COVID19")){
-
-      yn <- utils::askYesNo("Package COVID19: new version available. Update now?")
-      if(!is.na(yn)) if(yn)
-        update()
-
-    } else {
-
-      packageStartupMessage("...up to date.")
-
-    }
-
-  }
-
-  packageStartupMessage(
-    '
-    ================================================================\n
-    
-    IMPORTANT NOTICE: 
-    This is the development version of the COVID-19 Data Hub
-    
-    Download the stable release from CRAN:
-    install.packages("COVID19")
-    
-    ================================================================\n
-    '
-  )
-  
-}
-
-update <- function(){
-
-  detach("package:COVID19", unload=TRUE)
-  x <- try(remotes::install_github('covid19datahub/COVID19', quiet = FALSE, upgrade = FALSE), silent = TRUE)
-  library(COVID19)
-
-}
+# .onAttach <- function(libname, pkgname) {
+# 
+#   if(interactive() & requireNamespace('COVID19', quietly = TRUE)){
+# 
+#     packageStartupMessage("The coronavirus situation is changing fast. Checking for updates...")
+# 
+#     description <- try(readLines('https://raw.githubusercontent.com/covid19datahub/COVID19/master/DESCRIPTION'), silent = TRUE)
+#     if(class(description)=="try-error")
+#       return()
+#       
+#     id <- which(startsWith(prefix = "Version:", x = description))
+#     v  <- as.package_version(gsub(pattern = "^Version:\\s*", replacement = "", x = description[id]))
+# 
+#     if(v > utils::packageVersion(pkg = "COVID19")){
+# 
+#       yn <- utils::askYesNo("Package COVID19: new version available. Update now?")
+#       if(!is.na(yn)) if(yn)
+#         update()
+# 
+#     } else {
+# 
+#       packageStartupMessage("...up to date.")
+# 
+#     }
+# 
+#   }
+# 
+#   packageStartupMessage(
+#     '
+#     ================================================================\n
+#     
+#     IMPORTANT NOTICE: 
+#     This is the development version of the COVID-19 Data Hub
+#     
+#     Download the stable release from CRAN:
+#     install.packages("COVID19")
+#     
+#     ================================================================\n
+#     '
+#   )
+#   
+# }
+# 
+# update <- function(){
+# 
+#   detach("package:COVID19", unload=TRUE)
+#   x <- try(remotes::install_github('covid19datahub/COVID19', quiet = FALSE, upgrade = FALSE), silent = TRUE)
+#   library(COVID19)
+# 
+# }
 
 cachecall <- function(fun, ...){
   
@@ -167,7 +167,7 @@ cite <- function(x, src, verbose){
     
     dplyr::group_by(iso_alpha_3) %>%
     
-    dplyr::group_map(keep = TRUE, function(x, iso){
+    dplyr::group_map(.keep = TRUE, function(x, iso){
       
       iso   <- iso[[1]]
       level <- unique(x$administrative_area_level)
