+++ tests/testthat/test-today.R
@@ -11,7 +11,7 @@ test_that("today", {
       
       dplyr::group_by(iso_alpha_3) %>%
       
-      dplyr::group_map(keep = TRUE, function(x, iso){
+      dplyr::group_map(.keep = TRUE, function(x, iso){
        
         t <- is_equal(x[,cn], y[,cn])
         
