+++ index.md
@@ -407,8 +407,10 @@ Using the COVID-19 Data Hub? Open a [GitHub issue](https://github.com/covid19dat
 ## In the News
 
 - [Institute for Data Valorization, Canada](<https://ivado.ca/en/covid-19/#phares>)
+- [Winner of the eRum2020 CovidR contest](<https://milano-r.github.io/erum2020-covidr-contest/index.html>)
 - [Data Science and Economics, University of Milano](<https://dse.cdl.unimi.it/en/avviso/notice-detail/covid-data-analysis>)
 - [Top 50 R resources on Novel COVID-19 Coronavirus](<https://towardsdatascience.com/top-5-r-resources-on-covid-19-coronavirus-1d4c8df6d85f>)
+- [March 2020: "Top 40" New CRAN Packages](<https://rviews.rstudio.com/2020/04/27/march-2020-top-40-new-cran-packages/>)
 
 ## Authors
 
