+++ pkgdown.yml
@@ -204,7 +204,7 @@ articles:
   ZAF: iso/ZAF.html
   ZMB: iso/ZMB.html
   ZWE: iso/ZWE.html
-last_built: 2020-05-31T08:19Z
+last_built: 2020-05-31T12:21Z
 urls:
   reference: https://covid19datahub.io/reference
   article: https://covid19datahub.io/articles
