### 2020-06-01 b2ccfcf
New County Turkey
Data Source : https://ozanerturk.github.io/covid19-turkey-api/
