### 2020-06-01 07f6b37
Add new Country - Turkey
Data Source: https://github.com/ozanerturk/covid19-turkey-api/blob/master/dataset/timeline.csv
