+++ pkgdown.yml
@@ -204,7 +204,7 @@ articles:
   ZAF: iso/ZAF.html
   ZMB: iso/ZMB.html
   ZWE: iso/ZWE.html
-last_built: 2020-06-02T04:20Z
+last_built: 2020-06-02T08:22Z
 urls:
   reference: https://covid19datahub.io/reference
   article: https://covid19datahub.io/articles
