### 2020-06-04 2515e76
Add new country- TUR
Data Source: https://github.com/ozanerturk/covid19-turkey-api/blob/master/dataset/timeline.csv
