+++ R/covid19.R
@@ -43,7 +43,7 @@
 #' We have invested a lot of time and effort in creating \href{https://covid19datahub.io}{COVID-19 Data Hub}, please:
 #' 
 #' \itemize{
-#' \item cite \href{https://doi.org/10.13140/RG.2.2.11649.81763}{Guidotti and Ardia (2020)} in working papers and published papers that use it.
+#' \item cite \href{https://www.researchgate.net/publication/340771808_COVID-19_Data_Hub}{Guidotti and Ardia (2020)} in working papers and published papers that use it.
 #' \item place the URL \url{https://covid19datahub.io} in a footnote to help others find \href{https://covid19datahub.io}{COVID-19 Data Hub}.
 #' \item you assume full risk for the use of \href{https://covid19datahub.io}{COVID-19 Data Hub}. 
 #' We try our best to guarantee the data quality and consistency and the continuous filling of the Data Hub. 
