+++ R/iso_ITA.R
@@ -8,11 +8,20 @@ ITA <- function(level, cache){
   x <- pcmdpc_git(cache = cache, level = level)
 
   # id
-  if(level==2)
+  if(level==2){
+    
     x$id <- id(x$state, iso = "ITA", ds = "pcmdpc_git", level = level)
-  if(level==3)
+    
+  }
+  if(level==3){
+    
+    y <- ceeds_git(cache = cache) 
+    x <- merge(x, y, by = c("date", "city_code"), all = TRUE)
+    
     x$id <- id(x$city, iso = "ITA", ds = "pcmdpc_git", level = level)
-
+    
+  }
+  
   # return
   return(x)
 
