+++ pkgdown.yml
@@ -204,7 +204,7 @@ articles:
   ZAF: iso/ZAF.html
   ZMB: iso/ZMB.html
   ZWE: iso/ZWE.html
-last_built: 2020-06-06T08:22Z
+last_built: 2020-06-07T00:36Z
 urls:
   reference: https://covid19datahub.io/reference
   article: https://covid19datahub.io/articles
