### 2020-06-08 #2271 @will-rowe
Hi @HadrienG - would you be interested in reviewing this JOSS submission?
