+++ iso_TUR.R.R
@@ -1,0 +1,22 @@
+
+TUR <- function(level, cache){
+  # Author: Seehan
+  
+  # fallback
+  if(level>1)
+    return(NULL)
+  
+  # download
+  x <- covid19turkey(cache = cache, level = level) 
+  
+  # return
+  return(x)
+  
+}
+
+x <- covid19turkey(1, TRUE)
+
+
+require(COVID19)
+x <- covid19turkey(level = 1, TRUE)
+ds_check_format(x, level = 1)
