+++ R/ds_mzcr_cz.R
@@ -30,6 +30,7 @@ mzcr_cz <- function(level, cache){
     url <- sprintf("%s/osoby.csv", mzcr.covid.api)
     x   <- read.csv(url, cache = cache)
     
+    
     # regional
     x$date  <- as.Date(x[,1])
     x <- map_data(x, c(
@@ -47,6 +48,23 @@ mzcr_cz <- function(level, cache){
       dplyr::arrange(date) %>%
       dplyr::group_by(state) %>%
       dplyr::mutate(confirmed = cumsum(confirmed)) 
+  }
+  if(level == 3) {
+    # people confirmed/deaths/recovered (by districts)
+    url <- 'https://onemocneni-aktualne.mzcr.cz/api/v2/covid-19/kraj-okres-nakazeni-vyleceni-umrti.csv'
+    x   <- read.csv(url, cache = cache)
+    
+    x$date <- as.Date(x[,1])
+    x <- map_data(x, c(
+      "date",
+      "kraj_nuts_kod" = "state",
+      "okres_lau_kod" = "city", # rather district
+      "kumulativni_pocet_nakazenych" = "confirmed",
+      "kumulativni_pocet_vylecenych" = "recovered",
+      "kumulativni_pocet_umrti"      = "deaths"
+    ))
+    
+    print(x)
   }
   
   # return
