### 2020-06-15 4dac9e3
level 2 deaths and recovered added
