+++ R/ds_mzcr_cz.R
@@ -25,36 +25,13 @@ mzcr_cz <- function(level, cache){
     x <- merge(tests, confirmed, by = "date", all = TRUE)
 
   }
-  if(level == 2) {
-    # people confirmed (by regions)
-    url <- sprintf("%s/osoby.csv", mzcr.covid.api)
-    x   <- read.csv(url, cache = cache)
-    
-    
-    # regional
-    x$date  <- as.Date(x[,1])
-    x <- map_data(x, c(
-      "date",
-      "kraj" = "state"
-    ))
-
-    # bindings
-    date <- state <- NULL
-    
-    # cumulative
-    x <- x %>%
-      dplyr::group_by(date, state) %>%
-      dplyr::summarise(confirmed = dplyr::n()) %>%
-      dplyr::arrange(date) %>%
-      dplyr::group_by(state) %>%
-      dplyr::mutate(confirmed = cumsum(confirmed)) 
-  }
-  if(level == 3) {
+  # levels 2,3
+  if(level >= 2) {
     # people confirmed/deaths/recovered (by districts)
-    url <- 'https://onemocneni-aktualne.mzcr.cz/api/v2/covid-19/kraj-okres-nakazeni-vyleceni-umrti.csv'
-    x   <- read.csv(url, cache = cache)
-    
+    url    <- 'https://onemocneni-aktualne.mzcr.cz/api/v2/covid-19/kraj-okres-nakazeni-vyleceni-umrti.csv'
+    x      <- read.csv(url, cache = cache)
     x$date <- as.Date(x[,1])
+    
     x <- map_data(x, c(
       "date",
       "kraj_nuts_kod" = "state",
@@ -63,8 +40,19 @@ mzcr_cz <- function(level, cache){
       "kumulativni_pocet_vylecenych" = "recovered",
       "kumulativni_pocet_umrti"      = "deaths"
     ))
-    
-    print(x)
+  }
+  
+  if(level == 2) {
+    x <- x %>%
+      dplyr::group_by(date,state) %>%
+      dplyr::summarise(
+        confirmed = sum(confirmed),
+        recovered = sum(recovered),
+        deaths    = sum(deaths))
+  }
+  
+  if(level == 3) {
+   # nothing to do 
   }
   
   # return
