+++ R/ds_covid19turkey.R
@@ -1,0 +1,24 @@
+covid19turkey <- function(level, cache){
+  
+  # download
+  url <- "https://raw.githubusercontent.com/ozanerturk/covid19-turkey-api/master/dataset/timeline.csv"
+  x   <- read.csv(url, cache = cache) 
+  
+  # formatting
+  x <- map_data(x, c(
+    'date'               = 'date',
+    'totalTests'         = 'tests',
+    'totalCases'         = 'confirmed',
+    'totalDeaths'        = 'deaths',
+    'totalRecovered'     = 'recovered',
+    'totalIntensiveCare' = 'icu',
+    'totalIntubated'     = 'vent'
+  ))
+  
+  # filter date
+  x <- x[!is.na(x$date),]
+  x$date <- as.Date(x$date, format = "%d/%m/%Y")
+  
+  return(x)
+}
+
