### 2020-06-16 91f4390
Delete TUR.R
