+++ R/iso_TUR.R
@@ -1,12 +1,12 @@
 TUR <- function(level, cache){
-  # Author: Seehan
+  # Author: Muhammed Seehan Pengatkundil
   
   # fallback
   if(level>1)
     return(NULL)
   
   # download
-  x <- covid19turkey(cache = cache, level = level) 
+  x <- covid19turkey(level = level, cache = cache) 
   
   # return
   return(x)
