+++ R/ds_mzcr_cz.R
@@ -35,7 +35,7 @@ mzcr_cz <- function(level, cache){
     x <- map_data(x, c(
       "date",
       "kraj_nuts_kod" = "state",
-      "okres_lau_kod" = "city", # rather district
+      "okres_lau_kod" = "district",
       "kumulativni_pocet_nakazenych" = "confirmed",
       "kumulativni_pocet_vylecenych" = "recovered",
       "kumulativni_pocet_umrti"      = "deaths"
