### 2020-06-16 a3dff14
Update contributors.Rmd
