+++ vignettes/contributors.Rmd
@@ -7,7 +7,7 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-The following people are gratefully acknowledged for taking part in the data collection as a joint effort against COVID-19. Names are in alphabetical order, to view individual contributions click [here](https://github.com/covid19datahub/COVID19/graphs/contributors).
+The following people are gratefully acknowledged for taking part in the data collection as a joint effort against COVID-19. The names are in alphabetical order, to view individual contributions click [here](https://github.com/covid19datahub/COVID19/graphs/contributors).
 
 - [Baby Rijin](https://github.com/rijinbaby)
 - [Benes Martin](https://github.com/martinbenes1996)
