### 2020-06-18 #2271 @will-rowe
Hi @linuxscout and @MDebasish - are either of you able to review this submission to JOSS please?
