+++ pkgdown.yml
@@ -206,7 +206,7 @@ articles:
   ZAF: iso/ZAF.html
   ZMB: iso/ZMB.html
   ZWE: iso/ZWE.html
-last_built: 2020-06-20T12:24Z
+last_built: 2020-06-20T16:19Z
 urls:
   reference: https://covid19datahub.io/reference
   article: https://covid19datahub.io/articles
