+++ vignettes/doc/data.Rmd
@@ -7,7 +7,7 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-Download the data from the [here](/articles/data.html)
+Download the data [here](/articles/data.html)
 
 ## Identifiers
 
