### 2020-06-21 #2271 @linuxscout
Salam,
Where can I find the paper and the check list to do?
Thanks
