### 2020-06-21 #2271 @ArdiaD
thx @linuxscout for reviewing the paper and the data hub. The paper is here: https://github.com/covid19datahub/COVID19/tree/master/inst/joss => PDF
