### 2020-06-21 #2271 @ArdiaD
The checklist (at least the steps) are reported here: https://joss.readthedocs.io/en/latest/review_checklist.html#
I guess you'll have access to a real "check box" page with all these points mentioned. Am I correct @will-rowe ?
