+++ R/iso_USA.R
@@ -22,11 +22,17 @@ USA <- function(level, cache){
   }
   if(level==3){
     
-    # download
+    # JHU
     x <- jhucsse_git(file = "US", cache = cache, level = level, country = "USA")
-    
-    # id
     x$id <- id(x$id, iso = "USA", ds = "jhucsse_git", level = level)
+    
+    # NyTimes
+    y <- nytimes_git(cache = cache, level = level)
+    y$id <- id(y$fips, iso = "USA", ds = "nytimes_git", level = level)
+    
+    # combine
+    x <- x[!(x$fips %in% y$fips),]
+    x <- bind_rows(x, y)
     
   }
     
