+++ NEWS.md
@@ -1,0 +1,6 @@
+# COVID19 2.0.0 (2020-06-21)
+
+## Major changes
+
+* **US data** level 3 (counties). Change data provider from [JHU CSSE](https://github.com/CSSEGISandData/COVID-19) to [New York Times](https://github.com/nytimes/covid-19-data). Only minor counties not provided by [New York Times](https://github.com/nytimes/covid-19-data) are still from [JHU CSSE](https://github.com/CSSEGISandData/COVID-19).
+
