### 2020-06-21 #2376 @whedon
[REVIEW]: COVID-19 Data Hub
**Submitting author:** ArdiaD (<a href="http://orcid.org/0000-0003-2823-782X">David Ardia</a>)
**Repository:** <a href="https://github.com/covid19datahub/COVID19/" target ="_blank">https://github.com/covid19datahub/COVID19/</a>
**Version:** v2.0.3

- [x] **Contribution and authorship:** Has the submitting author (ArdiaD) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] **Contribution and authorship:** Has the submitting author (ArdiaD) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
