+++ pkgdown.yml
@@ -207,7 +207,7 @@ articles:
   ZMB: iso/ZMB.html
   ZWE: iso/ZWE.html
   usage: usage.html
-last_built: 2020-06-22T08:22Z
+last_built: 2020-06-22T12:27Z
 urls:
   reference: https://covid19datahub.io/reference
   article: https://covid19datahub.io/articles
