### 2020-06-24 c10b87f
add Japan data
