### 2020-06-24 6b753a6
add Japan data
