+++ pkgdown.yml
@@ -207,7 +207,7 @@ articles:
   ZMB: iso/ZMB.html
   ZWE: iso/ZWE.html
   usage: usage.html
-last_built: 2020-06-24T20:24Z
+last_built: 2020-06-25T00:38Z
 urls:
   reference: https://covid19datahub.io/reference
   article: https://covid19datahub.io/articles
