+++ pkgdown.yml
@@ -207,7 +207,7 @@ articles:
   ZMB: iso/ZMB.html
   ZWE: iso/ZWE.html
   usage: usage.html
-last_built: 2020-07-01T20:28Z
+last_built: 2020-07-02T00:45Z
 urls:
   reference: https://covid19datahub.io/reference
   article: https://covid19datahub.io/articles
