### 2020-07-04 #2376 @linuxscout
@will-rowe, I need some explanation from authors, on this feature!

>

>  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
