+++ vignettes/contributors.Rmd
@@ -22,6 +22,6 @@ The following people are gratefully acknowledged for taking part in the data col
 - [Pengatkundil Muhammed](https://github.com/muhammedseehan-commits)
 - [Yuan Zhou](https://github.com/yuanzhouIR)
 
-__Want to contribute?__ Join us on [Slack](https://join.slack.com/t/covid19datahub/shared_invite/zt-e921sryd-Sb97p~skvovQM6NuXFUNCw) and get started on [GitHub](articles/contributors.html)
+__Want to contribute?__ Join us on [Slack](https://join.slack.com/t/covid19datahub/shared_invite/zt-e921sryd-Sb97p~skvovQM6NuXFUNCw) and get started on [GitHub](https://github.com/covid19datahub/COVID19/)
 
 `r gsub("^# ", "## ", readr::read_file('../LICENSE.md'))`
