### 2020-07-05 #2376 @ArdiaD
> @will-rowe, I need some explanation from authors, on this feature!
>
> >
>
> > Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

@emanuele-guidotti what would you say here?
