### 2020-07-05 #2376 @ArdiaD
> Hey @linuxscout and @mbod - just checking in and seeing how the reviews are going. Let me know if you need anything!

thx @will-rowe for following up!
