### 2020-07-05 #2376 @ArdiaD
> @will-rowe, I need some explanation from authors, on this feature!
>
> >
>
> > Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Hi @linuxscout. We have implemented a set of unit testing functions in the package. This is available here: https://github.com/covid19datahub/COVID19/tree/master/tests/testthat

We follow the standard practice in R, using testthat.
