+++ pkgdown.yml
@@ -207,7 +207,7 @@ articles:
   ZMB: iso/ZMB.html
   ZWE: iso/ZWE.html
   usage: usage.html
-last_built: 2020-07-05T20:24Z
+last_built: 2020-07-06T00:45Z
 urls:
   reference: https://covid19datahub.io/reference
   article: https://covid19datahub.io/articles
