+++ pkgdown.yml
@@ -207,7 +207,7 @@ articles:
   ZMB: iso/ZMB.html
   ZWE: iso/ZWE.html
   usage: usage.html
-last_built: 2020-07-06T20:25Z
+last_built: 2020-07-07T00:49Z
 urls:
   reference: https://covid19datahub.io/reference
   article: https://covid19datahub.io/articles
