### 2020-07-09 covid19datahub/COVID19#90 @will-rowe
JOSS review
Hi there.

Thank you for the submission - this is a great resource! This is my review as part of https://github.com/openjournals/joss-reviews/issues/2376. Please can you address the following comments:

* tests / continuous integration

I can see the tests directory but I'm not entirely clear on how to run these. I don't code using R so this might be why. Regardless, could you please add some documentation for running these tests (or make it clearer where this documentation is if it already exists). Can you also add some continuous integration for your test suite please - travis or similar would be great.

* references

Please could you include DOIs for the references where you can.

* minor paper comments

Although the paper is very well written and the summary is nice, the actual software description/purpose isn't included until the final paragraph which is on the second page of the paper. In terms of readability/impact, maybe you could introduce the datahub earlier on? Feel free to ignore this comment if you wish. Secondly, and this might be intentional, there are a few extra full stops after first mention of Excel!
