### 2020-07-09 covid19datahub/COVID19#90 @ArdiaD
Hi @will-rowe thx a lot for the fast and detailed review. We will come back asap with our fixes!
