+++ inst/joss/paper.bib
@@ -13,7 +13,8 @@
   number={5},
   pages={533--534},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/S1473-3099(20)30120-1}
 }
 
 @article{hale2020oxford,
