### 2020-07-10 c7137e6
Adding travis build status to readme
This is not an essential change for JOSS, but makes it easier for users to check build status. Feel free to ignore the PR or to add the badge yourself!
