### 2020-07-10 #2376 @will-rowe
Great - thank you for your prompt responses @ArdiaD and @emanuele-guidotti

You've addressed my comments as a reviewer, and in addition to @linuxscout's review, I think we're ready to proceed to acceptance!

As a side note, I just suggested adding a build status badge (https://github.com/covid19datahub/COVID19/pull/91), but this is not required by JOSS.

@ArdiaD - please could you now make a tagged release and archive, and report the version number and archive DOI in this thread.
