### 2020-07-10 #2376 @kyleniemeyer
Hi @ArdiaD, we are nearly done here, but I have just a few things for you to fix in the paper:

- Please put commas between the lists on page 2 (both in the Data Collection section)
- please remove either "e.g." or "etc." in the "geographic information ..." bullet on page 2; you should only have one or the other of these

Thanks!
