### 2020-07-10 #2376 @kyleniemeyer
Congrats @ArdiaD and @emanuele-guidotti on your article's publication in JOSS!

Many thanks to @linuxscout and @will-rowe for reviewing this, and @will-rowe for editing!
